Initial commit
This contains a working version with some structure to add an authentication system with mocked calls to the backend.

It needs some improvements like adding a confirmation dialog when clicking on the reset game button, add a timer, make it so pressing enter on the login dialog makes login. And of course, adding a real backend. The idea of the backend is just keep track of the user's past games.

ps. Im commiting this months after I've coded it, so I might be missing some important details.